+++ src/scripts/articles/articleFormManager.js
